requirements/base.txt
# 웹
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.9

# DB
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# 검색
opensearch-py[async]>=2.4.0

# 문서 파싱
PyMuPDF>=1.24.0
ebooklib>=0.18
python-docx>=1.1.0
beautifulsoup4>=4.12.0
chardet>=5.2.0
pytesseract>=0.3.10
Pillow>=10.0.0

# 작업큐
celery[redis]>=5.3.0
redis>=5.0.0

# 유틸리티
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
watchdog>=4.0.0
httpx>=0.27.0
